Added bit operations, working on entry creation for split index k.

+++ src/index/split_index_1.hpp
@@ -31,7 +31,7 @@ protected:
     size_t calcEntrySizeB(const char *entry) const override;
 
     /** Returns the number of words (word parts) stored in [entry]. */
-    virtual int calcEntryNWords(const char *entry) const;
+    virtual size_t calcEntryNWords(const char *entry) const;
 
     /** Splits [word] into two and stores the parts (prefix and suffix) in prefixBuf and suffixBuf, resp. */
     void storePrefixSuffixInBuffers(const std::string &word);
